requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flake8==3.8.4
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
requests==2.25.1
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
